Safari-611.1.10/LayoutTests/canvas/philip/tests/2d.state.saverestore.shadowOffsetY.html

<!DOCTYPE html>
<title>Canvas test: 2d.state.saverestore.shadowOffsetY</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<body>
<p id="passtext">Pass</p>
<p id="failtext">Fail</p>
<p class="output">These images should be identical:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

// Test that restore() undoes any modifications
var old = ctx.shadowOffsetY;
ctx.save();
ctx.shadowOffsetY = 5;
ctx.restore();
_assertSame(ctx.shadowOffsetY, old, "ctx.shadowOffsetY", "old");

// Also test that save() doesn't modify the values
ctx.shadowOffsetY = 5;
old = ctx.shadowOffsetY;
    // we're not interested in failures caused by get(set(x)) != x (e.g.
    // from rounding), so compare against 'old' instead of against 5
ctx.save();
_assertSame(ctx.shadowOffsetY, old, "ctx.shadowOffsetY", "old");
ctx.restore();


});
</script>
