LayoutTests/canvas/philip/tests/2d.state.saverestore.shadowOffsetY-expected.txt

< [index] >
2d.state.saverestore.shadowOffsetY
save()/restore() works for shadowOffsetY
References: 2d.state.shadowOffsetY
Actual output:
Passed
